+++ integration/tests/common-extensions/computed.spec.ts
@@ -1,12 +1,12 @@
-/* eslint-disable */
 import { Injectable } from '@angular/core';
-import { NgxsModule, State, Store } from '@ngxs/store';
-import { DataAction, Computed, StateRepository } from '@ngxs-labs/data/decorators';
 import { TestBed } from '@angular/core/testing';
 import { NgxsDataPluginModule } from '@ngxs-labs/data';
+import { Computed, DataAction, StateRepository } from '@ngxs-labs/data/decorators';
 import { NgxsDataRepository, NgxsImmutableDataRepository } from '@ngxs-labs/data/repositories';
 import { NGXS_DATA_EXCEPTIONS } from '@ngxs-labs/data/tokens';
-import { getSequenceIdFromTarget } from '@ngxs-labs/data/internals';
+import { NgxsModule, State, Store } from '@ngxs/store';
+import { NgxsDataSequence } from '@ngxs-labs/data/internals';
+import { BehaviorSubject } from 'rxjs';
 
 describe('[TEST]: Computed fields', () => {
     it('should be throw when invalid annotated', () => {
@@ -74,8 +74,6 @@ describe('[TEST]: Computed fields', () => {
         expect(b.snapshot).toEqual('world');
         expect(b.snapshot).toEqual('world');
         expect(b.countSnapshot).toEqual(3);
-
-        expect(getSequenceIdFromTarget(b)).toEqual(2);
     });
 
     describe('calculate total', () => {
@@ -351,5 +349,237 @@ describe('[TEST]: Computed fields', () => {
         expect(b.snapshot).toEqual(4);
         expect(a.snapshot === a.getState()).toBeTruthy();
         expect(b.snapshot === b.getState()).toBeTruthy();
+    });
+
+    it('should be trigger when store changes', () => {
+        interface Model {
+            value: number;
+        }
+
+        @StateRepository()
+        @State({
+            name: 'b',
+            defaults: { value: 2 }
+        })
+        @Injectable()
+        class B extends NgxsDataRepository<Model> {}
+
+        @StateRepository()
+        @State({
+            name: 'a',
+            defaults: { value: 1 }
+        })
+        @Injectable()
+        class A extends NgxsDataRepository<Model> {
+            public heavyCount: number = 0;
+
+            constructor(private b: B) {
+                super();
+            }
+
+            @Computed()
+            public get sum(): number {
+                this.heavyCount++;
+                return this.snapshot.value + this.b.snapshot.value;
+            }
+        }
+
+        TestBed.configureTestingModule({
+            imports: [NgxsModule.forRoot([A, B]), NgxsDataPluginModule.forRoot()]
+        });
+
+        const store: Store = TestBed.get(Store);
+        const a: A = TestBed.get(A);
+        const b: B = TestBed.get(B);
+        const stream: NgxsDataSequence = TestBed.get(NgxsDataSequence);
+
+        expect(stream.sequenceValue).toEqual(1);
+        expect(store.snapshot()).toEqual({ b: { value: 2 }, a: { value: 1 } });
+
+        expect(a.snapshot).toEqual({ value: 1 });
+        expect(a.snapshot).toEqual(a.snapshot);
+        expect(a.snapshot).toEqual(a.snapshot);
+        expect(a.snapshot).toEqual(a.snapshot);
+
+        expect(b.snapshot).toEqual({ value: 2 });
+        expect(b.snapshot).toEqual(b.snapshot);
+        expect(b.snapshot).toEqual(b.snapshot);
+        expect(b.snapshot).toEqual(b.snapshot);
+
+        expect(a.sum).toEqual(3);
+        expect(a.sum).toEqual(3);
+        expect(a.sum).toEqual(3);
+        expect(a.sum).toEqual(3);
+        expect(a.sum).toEqual(3);
+        expect(a.sum).toEqual(3);
+        expect(a.sum).toEqual(3);
+
+        expect(a.heavyCount).toEqual(1);
+
+        store.reset({ a: { value: 5 }, b: { value: 10 } });
+
+        expect(stream.sequenceValue).toEqual(2);
+        expect(store.snapshot()).toEqual({ b: { value: 10 }, a: { value: 5 } });
+
+        expect(a.snapshot).toEqual({ value: 5 });
+        expect(a.snapshot).toEqual(a.snapshot);
+        expect(a.snapshot).toEqual(a.snapshot);
+        expect(a.snapshot).toEqual(a.snapshot);
+
+        expect(b.snapshot).toEqual({ value: 10 });
+        expect(b.snapshot).toEqual(b.snapshot);
+        expect(b.snapshot).toEqual(b.snapshot);
+        expect(b.snapshot).toEqual(b.snapshot);
+
+        expect(a.sum).toEqual(15);
+        expect(a.sum).toEqual(15);
+        expect(a.sum).toEqual(15);
+        expect(a.sum).toEqual(15);
+        expect(a.sum).toEqual(15);
+        expect(a.sum).toEqual(15);
+        expect(a.sum).toEqual(15);
+
+        expect(a.heavyCount).toEqual(2);
+
+        stream.ngOnDestroy();
+        store.reset({ a: { value: 0 }, b: { value: 0 } });
+        expect(stream.sequenceValue).toEqual(0);
+
+        store.reset({ a: { value: 1 }, b: { value: 1 } });
+        expect(stream.sequenceValue).toEqual(0);
+    });
+
+    it('does not recalculate when use state from third-party service', () => {
+        @Injectable()
+        class MyFirstCountService {
+            private values$: BehaviorSubject<number> = new BehaviorSubject(0);
+
+            public increment(): void {
+                this.values$.next(this.getValue() + 1);
+            }
+
+            public getValue(): number {
+                return this.values$.getValue();
+            }
+        }
+
+        @StateRepository()
+        @State({
+            name: 'count',
+            defaults: 0
+        })
+        @Injectable()
+        class MySecondCountState extends NgxsDataRepository<number> {
+            constructor(private readonly first: MyFirstCountService) {
+                super();
+            }
+
+            @Computed()
+            public get sum(): number {
+                return this.snapshot + this.first.getValue();
+            }
+
+            @DataAction()
+            public increment(): void {
+                this.ctx.setState((state: number) => ++state);
+            }
+        }
+
+        TestBed.configureTestingModule({
+            imports: [NgxsModule.forRoot([MySecondCountState]), NgxsDataPluginModule.forRoot()],
+            providers: [MyFirstCountService]
+        });
+
+        const first: MyFirstCountService = TestBed.get(MyFirstCountService);
+        const second: MySecondCountState = TestBed.get(MySecondCountState);
+
+        expect(first.getValue()).toEqual(0);
+        expect(second.snapshot).toEqual(0);
+        expect(second.sum).toEqual(0);
+
+        second.increment();
+        second.increment();
+
+        expect(first.getValue()).toEqual(0);
+        expect(second.snapshot).toEqual(2);
+        expect(second.sum).toEqual(2);
+
+        first.increment();
+        first.increment();
+        first.increment();
+
+        // Expect invalid behavior
+        expect(first.getValue()).toEqual(3);
+        expect(second.snapshot).toEqual(2);
+        expect(second.sum).toEqual(2);
+    });
+
+    it('recalculate sum when use state from third-party service', () => {
+        @Injectable()
+        class MyFirstCountService {
+            private values$: BehaviorSubject<number> = new BehaviorSubject(0);
+
+            constructor(private readonly sequence: NgxsDataSequence) {}
+
+            public increment(): void {
+                this.values$.next(this.getValue() + 1);
+                this.sequence.updateSequence();
+            }
+
+            public getValue(): number {
+                return this.values$.getValue();
+            }
+        }
+
+        // noinspection DuplicatedCode
+        @StateRepository()
+        @State({
+            name: 'count',
+            defaults: 0
+        })
+        @Injectable()
+        class MySecondCountState extends NgxsDataRepository<number> {
+            constructor(private readonly first: MyFirstCountService) {
+                super();
+            }
+
+            @Computed()
+            public get sum(): number {
+                return this.snapshot + this.first.getValue();
+            }
+
+            @DataAction()
+            public increment(): void {
+                this.ctx.setState((state: number) => ++state);
+            }
+        }
+
+        TestBed.configureTestingModule({
+            imports: [NgxsModule.forRoot([MySecondCountState]), NgxsDataPluginModule.forRoot()],
+            providers: [MyFirstCountService]
+        });
+
+        const first: MyFirstCountService = TestBed.get(MyFirstCountService);
+        const second: MySecondCountState = TestBed.get(MySecondCountState);
+
+        expect(first.getValue()).toEqual(0);
+        expect(second.snapshot).toEqual(0);
+        expect(second.sum).toEqual(0);
+
+        second.increment();
+        second.increment();
+
+        expect(first.getValue()).toEqual(0);
+        expect(second.snapshot).toEqual(2);
+        expect(second.sum).toEqual(2);
+
+        first.increment();
+        first.increment();
+        first.increment();
+
+        // Expect valid behavior
+        expect(first.getValue()).toEqual(3);
+        expect(second.snapshot).toEqual(2);
+        expect(second.sum).toEqual(5);
     });
 });