+++ lib/internals/src/utils/computed/get-computed-options.ts
@@ -1,7 +1,0 @@
-import { Any, ComputedOptions } from '@ngxs-labs/data/typings';
-
-import { computedKey } from '../common/computed-key';
-
-export function getComputedOptions(target: Any, key: string | symbol): ComputedOptions | null {
-    return target[computedKey(key)] ?? null;
-}