+++ lib/typings/src/public_api.ts
@@ -40,3 +40,4 @@ export { Descriptor } from './types/descriptor';
 export { DispatchedResult } from './types/dispatched-result';
 export { DataStateClass, StateClassDecorator, StateArgumentDecorator } from './types/data-state-class';
 export { ComputedOptions } from './types/computed-options';
+export { ComputedCacheMap } from './types/computed-cache-map';